requirements.txt
asgiref==3.3.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.3
Django==3.2a1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.3.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
mysqlclient==2.1.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
pygame==2.1.0
PyJWT==2.0.1
pytz==2021.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.9
virtualenv==20.10.0
webencodings==0.5.1
whitenoise==5.3.0
